v0.5.9.1: fix type of mutation probability

+++ moead_framework/core/genetic_operator/combinatorial/cross_mut.py
@@ -27,7 +27,7 @@ class CrossoverAndMutation(GeneticOperator):
         if kwargs.get("mutation_probability") is None:
             self.mutation_probability = 1 / (len(self.solutions[0]))
         else:
-            self.mutation_probability = int(kwargs.get("mutation_probability"))
+            self.mutation_probability = float(kwargs.get("mutation_probability"))
 
     def run(self):
         self.number_of_solution_is_correct(n=2)
